Midi is now run from the server. However, we don't get updates when mappings are changed in the webpage
